+++ tests/test_cryo_mut.rs
@@ -19,31 +19,30 @@ use std::{
 
 #[test]
 fn new() {
-    let mut var = 42;
-    cryo!(let _cryo: CryoMut<_> = &mut var);
+    with_cryo(&mut 42, |_| {});
 }
 
 #[test]
 fn read() {
-    let mut var = 42;
-    cryo!(let cryo_mut: CryoMut<_> = &mut var);
-    assert_eq!(*cryo_mut.read(), 42);
+    with_cryo(&mut 42, |cryo_mut| {
+        assert_eq!(*cryo_mut.read(), 42);
+    });
 }
 
 #[test]
 fn read2() {
-    let mut var = 42;
-    cryo!(let cryo_mut: CryoMut<_> = &mut var);
-    let b1 = cryo_mut.read();
-    let _b2 = cryo_mut.read();
-    assert_eq!(*b1, 42);
+    with_cryo(&mut 42, |cryo_mut| {
+        let b1 = cryo_mut.read();
+        let _b2 = cryo_mut.read();
+        assert_eq!(*b1, 42);
+    });
 }
 
 #[test]
 fn write() {
-    let mut var = 42;
-    cryo!(let cryo_mut: CryoMut<_> = &mut var);
-    assert_eq!(*cryo_mut.write(), 42);
+    with_cryo(&mut 42, |cryo_mut| {
+        assert_eq!(*cryo_mut.write(), 42);
+    });
 }
 
 #[test]
@@ -65,32 +64,32 @@ fn try_get_mut_fail() {
 
 #[test]
 fn block_on_drop() {
-    let mut var = 42;
-    cryo!(let cryo_mut: CryoMut<_, cryo::SyncLock> = &mut var);
-    let borrow = cryo_mut.read();
-    spawn(move || {
-        sleep(Duration::from_millis(50));
-        drop(borrow);
+    with_cryo((&mut 42, lock_ty::<SyncLock>()), |cryo_mut| {
+        let borrow = cryo_mut.read();
+        spawn(move || {
+            sleep(Duration::from_millis(50));
+            drop(borrow);
+        });
     });
 }
 
 #[test]
 fn block_by_exclusive_access() {
-    let mut var = 42;
-    cryo!(let cryo_mut: CryoMut<_, cryo::SyncLock> = &mut var);
-    let borrow = cryo_mut.read();
-    spawn(move || {
-        sleep(Duration::from_millis(100));
-        assert_eq!(*borrow, 42);
-        drop(borrow);
-    });
-    assert_eq!(std::mem::replace(&mut *cryo_mut.write(), 56), 42);
+    with_cryo((&mut 42, lock_ty::<SyncLock>()), |cryo_mut| {
+        let borrow = cryo_mut.read();
+        spawn(move || {
+            sleep(Duration::from_millis(100));
+            assert_eq!(*borrow, 42);
+            drop(borrow);
+        });
+        assert_eq!(std::mem::replace(&mut *cryo_mut.write(), 56), 42);
 
-    let mut borrow = cryo_mut.write();
-    spawn(move || {
-        sleep(Duration::from_millis(100));
-        assert_eq!(std::mem::replace(&mut *borrow, 72), 56);
-        drop(borrow);
+        let mut borrow = cryo_mut.write();
+        spawn(move || {
+            sleep(Duration::from_millis(100));
+            assert_eq!(std::mem::replace(&mut *borrow, 72), 56);
+            drop(borrow);
+        });
+        assert_eq!(std::mem::replace(&mut *cryo_mut.write(), 100), 72);
     });
-    assert_eq!(std::mem::replace(&mut *cryo_mut.write(), 100), 72);
 }